+++ R/p_value.R
@@ -171,6 +171,14 @@ p_value.rlm <- function(fit, ...) {
   make_it_so(p, se)
 }
 
+#' @export
+p_value.svyolr <- function(fit, ...) {
+  cs <- stats::coef(summary(fit))
+  p <- 2 * stats::pnorm(abs(cs[, 3]), lower.tail = FALSE)
+  se <- cs[, 2]
+
+  make_it_so(p, se)
+}
 
 #' @rdname p_value
 #' @export